Introduce libportal-qt5
Same changes as the previous commits, but needed some extra
parameters since Qt 5 uses C++.

Related: https://github.com/flatpak/libportal/issues/12
