+++ portal-test/qt5/meson.build
@@ -1,8 +1,6 @@
 
 add_languages('cpp', required : true)
 
-qt_dep = dependency('qt5', modules: ['Core', 'Gui', 'X11Extras', 'Widgets'])
-
 src = [
   'main.cpp',
   'portal-test-qt.h',
@@ -13,13 +11,13 @@ prep = qt.preprocess(
   moc_headers : 'portal-test-qt.h',
   moc_extra_arguments: ['-DMAKES_MY_MOC_HEADER_COMPILE'],
   ui_files : 'portal-test-qt.ui',
-  dependencies: qt_dep,
+  dependencies: libportal_qt5_dep,
 )
 
 executable('portal-test-qt',
   [src, prep],
   include_directories: [top_inc, libportal_inc],
-  dependencies: [qt_dep, libportal_dep],
+  dependencies: [libportal_qt5_dep],
   cpp_args : '-std=c++11',
   install : true,
 )
